Move odometry into its own class.
