+++ example code/positiontutorial/src/main/java/frc/robot/commands/AutonomousCommand.java
@@ -10,6 +10,7 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.subsystems.OctoDriveSubsystem;
+import frc.robot.subsystems.Odometry;
 import frc.robot.commands.MoveToCommand;
 import frc.robot.commands.TurnCommand;
 import edu.wpi.first.wpilibj.geometry.*;
@@ -28,24 +29,24 @@ public class AutonomousCommand extends CommandBase {
      * Autonomous mode constructor.  Sets the sequence of commands we want the robot to follow.
      * @param drive This is the Robot's drive.  Used to get odometry and set motor speeds.
      */
-    public AutonomousCommand(OctoDriveSubsystem drive ) {
+    public AutonomousCommand(OctoDriveSubsystem drive, Odometry odometry ) {
         // A sequence of commands to move the robot autonomously in a triangle.
         // The robot returns to its start position
         commandList = new SequentialCommandGroup(
             // Move to coordinate <1.0,0> (meters).  The robot starts at <0,0>, facing
             // the X axis, so this is essentially a move forward
-            new MoveToCommand( drive, new Translation2d( 1.0, 0.0 ) ),
+            new MoveToCommand( drive, odometry, new Translation2d( 1.0, 0.0 ) ),
             // Sharp Left turn to get us close to 90 degrees
-            new TurnCommand   ( drive, new Rotation2d(Math.toRadians(75.0)) ),
+            new TurnCommand  ( drive, odometry, new Rotation2d(Math.toRadians(75.0)) ),
             // The robot should now be at <1.0,0>, facing close to 90 degrees.  Going
             // forward now will move the robot along the Y axis.   
             // A "Move To" <1.0,0.50> command should move the robot forward about 50cm 
-            new MoveToCommand( drive, new Translation2d( 1.0, 0.5 )),  
+            new MoveToCommand( drive, odometry, new Translation2d( 1.0, 0.5 )),  
             // Another sharp turn.  This will turn the robot back toward where it started
             // (just over 180 degrees)          
-            new TurnCommand   ( drive, new Rotation2d(Math.toRadians( 190 )) ),      
+            new TurnCommand  ( drive, odometry, new Rotation2d(Math.toRadians( 190 )) ),      
             // Move back to the start.    
-            new MoveToCommand( drive, new Translation2d(0.0, 0.0 ))  
+            new MoveToCommand( drive, odometry, new Translation2d(0.0, 0.0 ))  
         );
     }
